+++ plugin/nodes/ShortText/basic/index.tsx
@@ -1,17 +1,19 @@
-import React from 'react';
+// import React from 'react';
+const React = window.React;
 
 interface IProps {
   handler: CFUIEventHandler;
-  node: CFMarkQuestion;
+  node: CFFillQuestion;
 }
 
-class Range extends React.Component<IProps> {
+class ShortTextBasic extends React.Component<IProps> {
   /**
    * 渲染组件
    */
   render(): JSX.Element {
     const { node, handler } = this.props;
     return <div>
+      <span>ShortTextBasic</span>
       {node.images.map(image => {
         return <img src={image.large}
           key={image.id}
@@ -24,20 +26,18 @@ class Range extends React.Component<IProps> {
       {node.options.map(opt => {
         return <div key={opt.renderId}>
           <span>{opt.text}</span>
-          <input type="range"
-            min={opt.minValue}
-            max={opt.maxValue}
-            step={opt.step}
+          <input type="text"
             value={opt.value}
             onChange={(e) => {
               const value = (e.target as HTMLInputElement).value;
-              handler.handleOptionInput(value, opt, node);
+              handler.handleOptionInput(value, opt, node)
             }} />
-          <span>{opt.value || 0}</span>
+          <span>{opt.value}</span>
+          <span style={{ background: 'red' }}>{opt.errorMessage}</span>
         </div>
       })}
     </div>
   }
 }
 
-export default Range;
+export default ShortTextBasic;